Move soon-to-be-deprecated waffle code to legacy module
This will be helpful in viewing the differences between the modules.
